making linux AppImages available locally instead of downloaded at build time
The last build failed because our python3 AppImage 404d on download. For buskill-app (the first repo that cloned *this* repo as a cross-platform GUI python app), I ended up making all dependencies download locally--well, just the ones outside apt that can't be downloaded securely.

 * https://github.com/BusKill/buskill-app/issues/2

For now this commit just copies some of that logic upstream to this repo so that the linux builds won't fail due to 404 in the future..